server/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
PySide6>=6.6.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
websockets>=12.0
qrcode[pil]>=7.4
psutil>=5.9.0
Pillow>=10.0.0
pyobjc-framework-Quartz>=10.0; sys_platform == "darwin"
